+++ src/types/query.d.ts
@@ -24,6 +24,8 @@ export module query {
   export function Drop(number: ExprArg, collection: ExprArg): Expr;
   export function Prepend(elements: ExprArg, collection: ExprArg): Expr;
   export function Append(elements: ExprArg, collection: ExprArg): Expr;
+  export function IsEmpty(collection: ExprArg): Expr;
+  export function IsNonEmpty(collection: ExprArg): Expr;
 
   export function Get(ref: ExprArg, ts?: ExprArg): Expr;
   export function KeyFromSecret(secret: ExprArg): Expr;